Update Check Palindrome.c

+++ Check Palindrome.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-    int n, reversedInteger = 0, remainder, originalInteger;
+    long long int n, reversedInteger = 0, remainder, originalInteger;
     printf("Enter an integer: ");
     scanf("%d", &n);
     originalInteger = n;
@@ -14,9 +14,9 @@ int main()
     }
     // palindrome if orignalInteger and reversedInteger are equal
     if (originalInteger == reversedInteger)
-        printf("%d is a palindrome.", originalInteger);
+        printf("%lld is a palindrome.", originalInteger);
     else
-        printf("%d is not a palindrome.", originalInteger);
+        printf("%lld is not a palindrome.", originalInteger);
     
     return 0;
 }
